integrator modified for initialize. default event manager seperated
