+++ pdip/integrator/base/__init__.py
@@ -1,3 +1,4 @@
 from .initializer import Initializer
 from .integrator import Integrator
-from .integrator_event_manager import IntegratorEventManager, DefaultIntegratorEventManager
+from .integrator_event_manager import IntegratorEventManager
+from .default_integrator_event_manager import DefaultIntegratorEventManager
