+++ pdip/integrator/pubsub/base/message_broker.py
@@ -1,4 +1,6 @@
 import multiprocessing
+from multiprocessing.managers import SyncManager
+from queue import Queue
 
 from .channel_queue import ChannelQueue
 from .event_listener import EventListener
@@ -8,14 +10,12 @@ from .message_broker_worker import MessageBrokerWorker
 class MessageBroker:
     def __init__(self, logger):
         self.logger = logger
-        self.manager = multiprocessing.Manager()
-        self.publish_queue = self.manager.Queue()
-        self.publish_channel = ChannelQueue(channel_queue=self.publish_queue)
-        self.message_queue = self.manager.Queue()
-        self.message_channel = ChannelQueue(channel_queue=self.message_queue)
-        self.worker: MessageBrokerWorker = MessageBrokerWorker(publish_channel=self.publish_channel,
-                                                               message_channel=self.message_channel,
-                                                               other_arg=None)
+        self.manager: SyncManager = None
+        self.publish_queue: Queue = None
+        self.publish_channel: ChannelQueue = None
+        self.message_queue: Queue = None
+        self.message_channel: ChannelQueue = None
+        self.worker: MessageBrokerWorker = None
         self.listener: EventListener = None
         self.subscribers = {}
         self.max_join_time = 60
@@ -31,15 +31,26 @@ class MessageBroker:
             self.worker.stop()
         if self.listener.is_alive() and not self.listener.stopped():
             self.listener.stop()
-        self.manager.shutdown()
+        if self.manager is not None:
+            self.manager.shutdown()
 
-    def get_publish_channel(self):
-        return self.publish_channel
+    def initialize(self):
+        self.manager = multiprocessing.Manager()
+        self.publish_queue = self.manager.Queue()
+        self.publish_channel = ChannelQueue(channel_queue=self.publish_queue)
+        self.message_queue = self.manager.Queue()
+        self.message_channel = ChannelQueue(channel_queue=self.message_queue)
+        self.worker: MessageBrokerWorker = MessageBrokerWorker(publish_channel=self.publish_channel,
+                                                               message_channel=self.message_channel,
+                                                               other_arg=None)
 
     def start(self):
         self.worker.start()
         self.listener = EventListener(channel=self.message_channel, subscribers=self.subscribers, logger=self.logger)
         self.listener.start()
+
+    def get_publish_channel(self):
+        return self.publish_channel
 
     def join(self):
         self.worker.join(self.max_join_time)
